wip: Quick and dirty handling of offsets
There's already elementary support for diffing pdfs with offsets in the
viewer. Especially with large PDFs this isn't yet useful as pixel
offset cannot be entered, it can just be adjusted through the
cursor ui elements. A first simple improvement is to double these
elements to provide a way to modify the offset in steps of 50.

In addition neither does the viewer support initial offset, nor does
the diff exporter support offset at all. Add support for OFFSET_X and
OFFSET_Y environment variables to improve those two use cases.

These changes are not meant to be merged upstream, they're rather just
a quick and dirty hack to help me diffing a couple of large files.
A proper solution would add command line parameters instead of
environment variables, and for example a dialog to enter offsets in the
ui.